Module Vector_db.Vec

module Float_array : sig ... end

type vector = Float_array.t

val compare_vector : vector -> vector -> int

val bin_shape_vector : Bin_prot.Shape.t

val bin_size_vector : vector Bin_prot.Size.sizer

val bin_write_vector : vector Bin_prot.Write.writer

val bin_writer_vector : vector Bin_prot.Type_class.writer

val bin_read_vector : vector Bin_prot.Read.reader

val __bin_read_vector__ : (int -> vector) Bin_prot.Read.reader

val bin_reader_vector : vector Bin_prot.Type_class.reader

val bin_vector : vector Bin_prot.Type_class.t

val sexp_of_vector : vector -> Sexplib0.Sexp.t

val vector_of_sexp : Sexplib0.Sexp.t -> vector

type t = {

id : string;

vector : vector;

}

this data type holds the vector representation of the underlying document and the id field is the file path location for the doc that the vecctor represents

include Ppx_compare_lib.Comparable.S with type t := t

val compare : t -> t -> int

include Bin_prot.Binable.S with type t := t

include Bin_prot.Binable.S_only_functions with type t := t

val bin_size_t : t Bin_prot.Size.sizer

val bin_write_t : t Bin_prot.Write.writer

val bin_read_t : t Bin_prot.Read.reader

val __bin_read_t__ : (int -> t) Bin_prot.Read.reader

This function only needs implementation if t exposed to be a polymorphic variant. Despite what the type reads, this does *not* produce a function after reading; instead it takes the constructor tag (int) before reading and reads the rest of the variant t afterwards.

val bin_shape_t : Bin_prot.Shape.t

val bin_writer_t : t Bin_prot.Type_class.writer

val bin_reader_t : t Bin_prot.Type_class.reader

val bin_t : t Bin_prot.Type_class.t

include Sexplib0.Sexpable.S with type t := t

val t_of_sexp : Sexplib0.Sexp.t -> t

val sexp_of_t : t -> Sexplib0.Sexp.t

module Io : sig ... end

val write_vectors_to_disk : t array -> path -> unit

Writes an array of vectors to disk using the Io.File module

parameter vectors
The array of vectors to be written to disk

parameter label
The label used as the file name for the output file

val read_vectors_from_disk : path -> t array

Reads an array of vectors from disk using the Io.File module

parameter label
The label used as the file name for the input file

returns
The array of vectors read from the file